+++ packages/orq-rc/src/orq_ai_sdk/models/conversationresponse.py
@@ -104,7 +104,7 @@ class ConversationResponse(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kfj3rrmj16pzxfv3315dvsan"
+        "conv_01kfjan6het575afkdwawqs3ve"
     )
     r"""Unique conversation identifier with `conv_` prefix."""