+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetitemop.py
@@ -1434,7 +1434,7 @@ class Evaluations3(BaseModel):
 
     source: Optional[CreateDatasetItemEvaluationsDatasetsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-01-22T05:45:22.258Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-01-22T07:45:45.157Z")
     r"""The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1502,7 +1502,7 @@ class Evaluations2(BaseModel):
 
     source: Optional[CreateDatasetItemEvaluationsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-01-22T05:45:22.257Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-01-22T07:45:45.157Z")
     r"""The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1570,7 +1570,7 @@ class Evaluations1(BaseModel):
 
     source: Optional[EvaluationsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-01-22T05:45:22.257Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-01-22T07:45:45.156Z")
     r"""The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1665,7 +1665,7 @@ class CreateDatasetItemResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-01-22T05:45:08.395Z")
+    updated: Optional[datetime] = parse_datetime("2026-01-22T07:45:31.576Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")