+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFJ3RSEMX5GJ6DZG87TVA014"
+        "01KFJAN72J6104V5F3JPW59H3V"
     )
     r"""The unique identifier of the data source"""