+++ packages/orq-rc/src/orq_ai_sdk/models/filegetop.py
@@ -64,7 +64,7 @@ class FileGetResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2026-01-22T05:45:11.893Z")
+    created: Optional[datetime] = parse_datetime("2026-01-22T07:45:35.020Z")
     r"""The date and time the resource was created"""
 
     @model_serializer(mode="wrap")