+++ packages/orq-rc/src/orq_ai_sdk/models/listcontactsop.py
@@ -208,7 +208,7 @@ class ListContactsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-01-22T05:45:08.395Z")
+    updated: Optional[datetime] = parse_datetime("2026-01-22T07:45:31.576Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")