+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFJ3RSEJJZG2JT2V3K6T3F01"
+        "01KFJAN72EDXSTS7KAREK4YN7S"
     )
     r"""The unique identifier of the data source"""