+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFJ3RSEKBR20DDVTHNMTCJZ9"
+        "01KFJAN72GBZRDE508MV5MBVJA"
     )
     r"""The unique identifier of the data source"""