+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFJ3RSEPC6RKYZ1AMNN8JKC7"
+        "01KFJAN72NNNHBXJ2QQZPSREZJ"
     )
     r"""The unique identifier of the data source"""